backend/requirements/prod.txt
Django==3.0.7
djangorestframework==3.12.1
django-cors-headers==3.7.0